Criação do Rest das telas de comando da aplicação. Ajustes nos models para dar suporte ao rest. Criação do código para verificar se o lançamento inicial foi efetuado.

+++ ErpWpf/Vendas/Component/View/Telas/LancamentoInicialView.xaml.cs
@@ -1,6 +1,9 @@
-﻿using System.Diagnostics;
+﻿using System;
+using System.Diagnostics;
 using System.Windows;
-using System.Windows.Input;
+using Erp.Business.Entity.Vendas.MovimentacaoCaixa.SubClass.LancamentoInicial;
+using Util;
+using Vendas.Properties;
 using Vendas.ViewModel.Forms;
 
 namespace Vendas.Component.View.Telas
@@ -19,31 +22,45 @@ namespace Vendas.Component.View.Telas
         {
             InitializeComponent();
             TxtValor.Focus();
-            Model.LancamentoEfetuado += Model_LancamentoEfetuado;
+            RestCommands.DataContext = DataContext;
             
         }
 
-        void Model_LancamentoEfetuado(object sender, System.EventArgs e)
+        protected override void OnPropertyChanged(DependencyPropertyChangedEventArgs e)
         {
-            Hide();
-        }
-
-        private void CmdCancelar_OnClick(object sender, RoutedEventArgs e)
-        {
-            Process.GetCurrentProcess().Kill();
-        }
-
-
-        private void LancamentoInicialView_OnPreviewKeyDown(object sender, KeyEventArgs e)
-        {
-            if (e.Key == Key.Escape)
+            base.OnPropertyChanged(e);
+            if (e.Property.Name == "DataContext")
             {
-                Hide();
+                Model.Sair += ModelOnSair;
             }
         }
 
-        private void CmdOk_OnClick(object sender, RoutedEventArgs e)
+        private void ModelOnSair(object sender, EventArgs eventArgs)
         {
+            Close();
+        }
+
+        
+        
+
+        //private void LancamentoInicialView_OnPreviewKeyDown(object sender, KeyEventArgs e)
+        //{
+        //    if (e.Key == Key.Escape)
+        //    {
+        //        Hide();
+        //    }
+        //}
+
+
+        private void LancamentoInicialView_OnClosed(object sender, EventArgs e)
+        {
+            var dia = LancamentoInicialRepository.DiaLancado(Settings.Default.Caixa, DateTime.Now, App.Proprietaria);
+            if (dia == null)
+            {
+                CustomMessageBox.MensagemInformativa("Não foi identificado um lançamento inicial de caixa.\n\n " +
+                                                     "A aplicação será encerrada.");
+                Process.GetCurrentProcess().Kill();
+            }
         }
     }
 }
